refactor(scan): split scan root owner
